Move the `Text` component into a submodule
I want to introduce a submodule for alignment, but then I'll need to make the
`text` module public, which will expose the public `Text` component definition,
resulting in it appearing in two different modules. To avoid confusion, I've
moved it to a new `text::component` submodule, only visible to super, like we
have with `stack::horizontal` and `stack::vertical`.

I'm not at all sold on the name `component`, but I can't think of anything
better at the moment -- I thought of using the component name, but `text::text`
seemed more confusing than helpful (and Clippy warns about it).
