+++ intuitive/src/components/mod.rs
@@ -212,7 +212,7 @@ pub use self::{
   empty::Empty,
   section::Section,
   stack::{horizontal::Stack as HStack, vertical::Stack as VStack},
-  text::Text,
+  text::component::Text,
 };
 use crate::element::Any as AnyElement;
